chore: moved listing data getter to NFTAuction utils
